Rename parser tests and remove old ones
